green: isolate fullscreen dock regression coverage
